Move GetVP, VerifyVP, Todo
